Phase B-5: kx-capability lib.rs split (Rule 3, behavior-preserving; D24 seam)
Fifth Phase B refactor per Rule 3. **Rule 1 max-strict surface** (D24
broker seam + D38 §1 token + D38 §2a probe). Pure move; zero functional
change.

Before: src/lib.rs 1304 lines (idempotency_token_for + EffectRequest +
BrokerHandle + CapabilityFailureReason + BrokerError + Capability trait
+ CapabilityBroker trait + LocalCapabilityBroker impl + inline
architectural review comment + 12 inline tests).

After: thin src/lib.rs 110 lines + 7 sibling modules:
- token.rs (D38 §1 idempotency_token_for helper)
- request.rs (EffectRequest + BrokerHandle)
- errors.rs (CapabilityFailureReason + BrokerError)
- capability.rs (Capability trait — load-bearing per D24)
- broker.rs (CapabilityBroker trait — load-bearing per D24)
- local.rs (LocalCapabilityBroker impl + inline architectural review)
- tests.rs (the 12 inline unit tests)

Re-exports preserve every previously-public path (kx_capability::{
idempotency_token_for, EffectRequest, BrokerHandle,
CapabilityFailureReason, BrokerError, Capability, CapabilityBroker,
LocalCapabilityBroker}).

Two minimal scope-widenings needed (Rule 1 max-strict surface — both
documented inline):

1. LocalCapabilityBroker::store field visibility: `S` → `pub(crate) S`.
   The inline tests reach into `broker.store.list_refs()` for
   content-store-side assertions (legacy from when tests lived inside the
   same module). The smallest valid scope is pub(crate); no public API
   surface change. A follow-up cleanup PR may introduce a typed accessor
   (`pub(crate) fn store(&self) -> &S`) if other patterns emerge.
2. tests.rs imports: added `WarrantField` to the kx_warrant import line
   (previously brought in via `super::*` from old lib.rs's `use
   kx_warrant::{..., WarrantField, ...};`). Pure import addition.

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-capability: clean
- cargo test --workspace --all-features: 645 + 8 (IDENTICAL)
- cargo test -p kx-capability: 12 + 4 + 4 + 1 = 21 (IDENTICAL)
- cargo clippy -p kx-capability --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean

Next Phase B slice: kx-projection (1305 lines; max-strict per Rule 1 —
9-cell recovery fold).
